requirements.txt
openai
anthropic
python-dotenv
rich
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
